Squash migrations and enrich API responses with related data
Consolidate migrations 001-007 into a single initial schema to simplify
fresh deployments. Enrich list/get endpoints for requests, responses,
and users with JOINed metadata (vote totals, comment counts, author
names, topics, composite scores, user's own vote). Add MaybeAuthUser
extractor for optional auth context, site stats endpoint, slug
validation, topic filtering, period/sort filters for requests, and
missing_criterion filter for responses. Extract composite score SQL
into a shared constant to reduce duplication. Move user stats into the
user profile endpoint and remove the separate leaderboard user stats
route.

+++ src/routes/comments.rs
@@ -137,7 +137,7 @@ async fn list_comments_inner(
         "SELECT c.id, c.author_id, u.display_name as author_display_name,
                 u.avatar_url as author_avatar_url, c.parent_id, c.depth,
                 c.body,
-                COALESCE(SUM(cv.value::bigint), 0) as vote_total,
+                COALESCE(SUM(cv.value::bigint), 0)::bigint as vote_total,
                 c.created_at, c.updated_at
          FROM comments c
          JOIN users u ON u.id = c.author_id
